+++ config/settings/test.py
@@ -97,6 +97,10 @@ PROJECTROLES_ALLOW_LOCAL_USERS = True
 
 # Samplesheets app settings
 SHEETS_ENABLE_CACHE = False  # Temporarily disabled to fix CI, see issue #556
+SHEETS_EXTERNAL_LINK_PATH = os.path.join(
+    ROOT_DIR, 'samplesheets/tests/config/ext_links.json'
+)
+
 
 # iRODS settings shared by iRODS using apps
 ENABLE_IRODS = True